Sniper Monkey
Created a Scriptable Object for the new Sniper Monkey
Created a Prefab for the Sniper Monkey using assets provided to you in the setup branch.
Used this Prefab in the Scriptable Object of the monkey
Added the Sniper Monkey Scriptable Object in the list of Monkey SOs being used in the game
Tweaked the parameters of the sniper monkey in its Scriptable Object.